Revert "insights: return pie chart insights with their own presentation options (#27237)" (#27302)
This reverts commit 8effcd4f40dc7680a28596e7e77390cc222a445e.

+++ internal/insights/insights.go
@@ -276,7 +276,7 @@ type LangStatsInsight struct {
 	ID             string
 	Title          string
 	Repository     string
-	OtherThreshold float64
+	OtherThreshold float32
 	OrgID          *int32
 	UserID         *int32
 }
